Crate uefi_raw

Raw interface for working with UEFI.

This crate is intended for implementing UEFI services. It is also used for implementing the uefi crate, which provides a safe wrapper around UEFI.

For creating UEFI applications and drivers, consider using the uefi crate instead of uefi-raw.

Modules

capsule

UEFI update capsules.

firmware_storage

Types related to firmware storage.

protocol

Protocol definitions.

table

Standard UEFI tables.

time

Date and time types.

Macros

guid

Create a Guid from a string at compile time.

newtype_enum

Interface a C-style enum as an integer newtype.

Structs

Guid

Globally-unique identifier.

Ipv4Address

An IPv4 internet protocol address.

Ipv6Address

An IPv6 internet protocol address.

MacAddress

A Media Access Control (MAC) address.

Status

UEFI uses status codes in order to report successes, errors, and warnings.

Type Aliases

Char8

One-byte character.

Char16

Two-byte character.

Event

Handle to an event structure.

Handle

Handle to a UEFI entity (protocol, image, etc).

PhysicalAddress

Physical memory address. This is always a 64-bit value, regardless of target platform.

VirtualAddress

Virtual memory address. This is always a 64-bit value, regardless of target platform.

Unions

IpAddress

An IPv4 or IPv6 internet protocol address.