uefi_raw

Module capsule

UEFI update capsules.

Capsules are used to pass information to the firmware, for example to trigger a firmware update.

Structs

CapsuleBlockDescriptor

Descriptor that defines a scatter-gather list for passing a set of capsules to the firmware.

CapsuleFlags

Capsule update flags.

CapsuleHeader

Common header at the start of a capsule.