uefi_raw::firmware_storage

Struct FirmwareVolumeAttributes

pub struct FirmwareVolumeAttributes(/* private fields */);

Corresponds to the C type EFI_FVB_ATTRIBUTES_2.

Implementations

impl FirmwareVolumeAttributes

pub const READ_DISABLED_CAP: Self

pub const READ_ENABLED_CAP: Self

pub const READ_STATUS: Self

pub const WRITE_DISABLED_CAP: Self

pub const WRITE_ENABLED_CAP: Self

pub const WRITE_STATUS: Self

pub const LOCK_CAP: Self

pub const LOCK_STATUS: Self

pub const STICKY_WRITE: Self

pub const MEMORY_MAPPED: Self

pub const ERASE_POLARITY: Self

pub const READ_LOCK_CAP: Self

pub const READ_LOCK_STATUS: Self

pub const WRITE_LOCK_CAP: Self

pub const WRITE_LOCK_STATUS: Self

pub const ALIGNMENT: Self

pub const WEAK_ALIGNMENT: Self

pub const ALIGNMENT_1: Self

pub const ALIGNMENT_2: Self

pub const ALIGNMENT_4: Self

pub const ALIGNMENT_8: Self

pub const ALIGNMENT_16: Self

pub const ALIGNMENT_32: Self

pub const ALIGNMENT_64: Self

pub const ALIGNMENT_128: Self

pub const ALIGNMENT_256: Self

pub const ALIGNMENT_512: Self

pub const ALIGNMENT_1K: Self

pub const ALIGNMENT_2K: Self

pub const ALIGNMENT_4K: Self

pub const ALIGNMENT_8K: Self

pub const ALIGNMENT_16K: Self

pub const ALIGNMENT_32K: Self

pub const ALIGNMENT_64K: Self

pub const ALIGNMENT_128K: Self

pub const ALIGNMENT_256K: Self

pub const ALIGNMENT_512K: Self

pub const ALIGNMENT_1M: Self

pub const ALIGNMENT_2M: Self

pub const ALIGNMENT_4M: Self

pub const ALIGNMENT_8M: Self

pub const ALIGNMENT_16M: Self

pub const ALIGNMENT_32M: Self

pub const ALIGNMENT_64M: Self

pub const ALIGNMENT_128M: Self

pub const ALIGNMENT_256M: Self

pub const ALIGNMENT_512M: Self

pub const ALIGNMENT_1G: Self

pub const ALIGNMENT_2G: Self

impl FirmwareVolumeAttributes

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u32

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl FirmwareVolumeAttributes

pub const fn iter(&self) -> Iter<FirmwareVolumeAttributes>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<FirmwareVolumeAttributes>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.