uefi_raw::protocol

Module device_path

The UEFI device path protocol, i.e., UEFI device paths.

This module provides (generated) ABI-compatible bindings to all known device path node types.

Terminology: Device Paths, Device Path Instances, and Device Path Nodes

An open UEFI device path protocol, also called device path, is a flexible and structured sequence of binary nodes that describes a route from the UEFI root to a particular device, controller, or file.

An entire device path can be made up of multiple device path instances, and each instance is made up of multiple device path nodes. A device path may contain multiple device-path instances separated by END_INSTANCE nodes, but typical paths contain only a single instance (in which case no END_INSTANCE node is needed). The entire device path is terminated with an END_ENTIRE node.

Each node represents a step in the path: PCI device, partition, filesystem, file path, etc. Each node represents a step in the path: PCI device, partition, filesystem, file path, etc.

Example of what a device path containing two instances (each comprised of three nodes) might look like:

┌──────┬──────┬──────────────╥───────┬──────────┬────────────┐
│ ACPI │ PCI  │ END_INSTANCE ║ CDROM │ FILEPATH │ END_ENTIRE │
└──────┴──────┴──────────────╨───────┴──────────┴────────────┘
↑      ↑      ↑              ↑       ↑          ↑            ↑
├─Node─╨─Node─╨─────Node─────╨─Node──╨───Node───╨────Node────┤
↑                            ↑                               ↑
├─── DevicePathInstance ─────╨────── DevicePathInstance ─────┤
│                                                            │
└──────────────────── Entire DevicePath ─────────────────────┘

Modules

acpi

Device path nodes for DeviceType::ACPI.

bios_boot_spec

Device path nodes for DeviceType::BIOS_BOOT_SPEC.

end

Device path nodes for DeviceType::END.

hardware

Device path nodes for DeviceType::HARDWARE.

media

Device path nodes for DeviceType::MEDIA.

messaging

Device path nodes for DeviceType::MESSAGING.

Structs

DevicePathFromTextProtocol

DevicePathProtocol

Device path protocol.

DevicePathToTextProtocol

DevicePathUtilitiesProtocol

DeviceSubType

Sub-type identifier for a device path node.

DeviceType

Type identifier for a device path node.