uefi_raw::protocol::file_system

Struct FileInfo

#[repr(C)]
pub struct FileInfo {
    pub size: u64,
    pub file_size: u64,
    pub physical_size: u64,
    pub create_time: Time,
    pub last_access_time: Time,
    pub modification_time: Time,
    pub attribute: FileAttribute,
    pub file_name: [Char16; 0],
}

Fields
size: u64file_size: u64physical_size: u64create_time: Timelast_access_time: Timemodification_time: Timeattribute: FileAttributefile_name: [Char16; 0]

The null-terminated name of the file. For a root directory, this is an empty string.

Note that this field is actually a variable-length array. In order to avoid making this struct a DST, the field is represented as a zero-length array here.

Implementations

impl FileInfo

pub const ID: Guid