uefi_raw::protocol::hii::config

Struct BrowserAction

#[repr(transparent)]
pub struct BrowserAction(pub usize);

Type of action taken by the form browser

Tuple Fields
0: usize
Implementations

impl BrowserAction

pub const CHANGING: BrowserAction

Called before the browser changes the value in the display (for questions which have a value) or takes an action (in the case of an action button or cross-reference). If EFI_SUCCESS is returned, the browser uses the value returned by Callback().

pub const CHANGED: BrowserAction

Called after the browser has changed its internal copy of the question value and displayed it (if appropriate). For action buttons, this is called after processing. Errors are ignored.

pub const RETRIEVE: BrowserAction

Called after the browser has read the current question value but before displaying it. If EFI_SUCCESS is returned, the updated value is used.

pub const FORM_OPEN: BrowserAction

Called for each question on a form prior to its value being retrieved or displayed. If a question appears on more than one form, this may be called more than once.

pub const FORM_CLOSE: BrowserAction

Called for each question on a form after processing any submit actions for that form. If a question appears on multiple forms, this will be called more than once.

pub const SUBMITTED: BrowserAction

Called after the browser submits the modified question value. ActionRequest is ignored.

pub const DEFAULT_STANDARD: BrowserAction

Represents the standard default action, selecting a default value based on lower-priority methods.

pub const DEFAULT_MANUFACTURING: BrowserAction

Represents the manufacturing default action, selecting a default value relevant to manufacturing.

pub const DEFAULT_SAFE: BrowserAction

Represents the safe default action, selecting the safest possible default value.

pub const DEFAULT_PLATFORM: BrowserAction

Represents platform-defined default values within a range of possible store identifiers.

pub const DEFAULT_HARDWARE: BrowserAction

Represents hardware-defined default values within a range of possible store identifiers.

pub const DEFAULT_FIRMWARE: BrowserAction

Represents firmware-defined default values within a range of possible store identifiers.