uefi_raw::protocol::network

Module pxe

Structs

PxeBaseCodeArpEntry

PxeBaseCodeBootType

PxeBaseCodeDhcpV4Packet

PxeBaseCodeDhcpV6Packet

PxeBaseCodeDiscoverInfo

PxeBaseCodeIcmpError

PxeBaseCodeIcmpErrorEcho

In the C API, this is an anonymous struct inside the definition of EFI_PXE_BASE_CODE_ICMP_ERROR.

PxeBaseCodeIpFilter

PxeBaseCodeIpFilterFlags

IP receive filters.

PxeBaseCodeMode

PxeBaseCodeMtftpInfo

PxeBaseCodeProtocol

PxeBaseCodeRouteEntry

PxeBaseCodeSrvlist

PxeBaseCodeTftpError

PxeBaseCodeTftpOpcode

PxeBaseCodeUdpOpFlags

Flags for UDP read and write operations.

Type Aliases

PxeBaseCodeUdpPort

Unions

PxeBaseCodeIcmpErrorUnion

In the C API, this is an anonymous union inside the definition of EFI_PXE_BASE_CODE_ICMP_ERROR.

PxeBaseCodePacket