uefi_raw::protocol::network::snp

Struct NetworkStatistics

#[repr(C)]
pub struct NetworkStatistics {

    pub rx_total_frames: u64,
    pub rx_good_frames: u64,
    pub rx_undersize_frames: u64,
    pub rx_oversize_frames: u64,
    pub rx_dropped_frames: u64,
    pub rx_unicast_frames: u64,
    pub rx_broadcast_frames: u64,
    pub rx_multicast_frames: u64,
    pub rx_crc_error_frames: u64,
    pub rx_total_bytes: u64,
    pub tx_total_frames: u64,
    pub tx_good_frames: u64,
    pub tx_undersize_frames: u64,
    pub tx_oversize_frames: u64,
    pub tx_dropped_frames: u64,
    pub tx_unicast_frames: u64,
    pub tx_broadcast_frames: u64,
    pub tx_multicast_frames: u64,
    pub tx_crc_error_frames: u64,
    pub tx_total_bytes: u64,
    pub collisions: u64,
    pub unsupported_protocol: u64,
    pub rx_duplicated_frames: u64,
    pub rx_decrypt_error_frames: u64,
    pub tx_error_frames: u64,
    pub tx_retry_frames: u64,

}

Network Statistics

The description of statistics on the network with the SNP’s statistics function is returned in this structure

Any of these statistics may or may not be available on the device. So, all the retriever functions of the statistics return None when a statistic is not supported

Fields
rx_total_frames: u64rx_good_frames: u64rx_undersize_frames: u64rx_oversize_frames: u64rx_dropped_frames: u64rx_unicast_frames: u64rx_broadcast_frames: u64rx_multicast_frames: u64rx_crc_error_frames: u64rx_total_bytes: u64tx_total_frames: u64tx_good_frames: u64tx_undersize_frames: u64tx_oversize_frames: u64tx_dropped_frames: u64tx_unicast_frames: u64tx_broadcast_frames: u64tx_multicast_frames: u64tx_crc_error_frames: u64tx_total_bytes: u64collisions: u64unsupported_protocol: u64rx_duplicated_frames: u64rx_decrypt_error_frames: u64tx_error_frames: u64tx_retry_frames: u64
Implementations

impl NetworkStatistics

pub const fn rx_total_frames(&self) -> Option<u64>

The total number of frames received, including error frames and dropped frames

pub const fn rx_good_frames(&self) -> Option<u64>

The total number of good frames received and copied into receive buffers

pub const fn rx_undersize_frames(&self) -> Option<u64>

The number of frames below the minimum length for the communications device

pub const fn rx_oversize_frames(&self) -> Option<u64>

The number of frames longer than the maximum length for the communications length device

pub const fn rx_dropped_frames(&self) -> Option<u64>

The number of valid frames that were dropped because the receive buffers were full

pub const fn rx_unicast_frames(&self) -> Option<u64>

The number of valid unicast frames received and not dropped

pub const fn rx_broadcast_frames(&self) -> Option<u64>

The number of valid broadcast frames received and not dropped

pub const fn rx_multicast_frames(&self) -> Option<u64>

The number of valid multicast frames received and not dropped

pub const fn rx_crc_error_frames(&self) -> Option<u64>

Number of frames with CRC or alignment errors

pub const fn rx_total_bytes(&self) -> Option<u64>

The total number of bytes received including frames with errors and dropped frames

pub const fn tx_total_frames(&self) -> Option<u64>

The total number of frames transmitted including frames with errors and dropped frames

pub const fn tx_good_frames(&self) -> Option<u64>

The total number of valid frames transmitted and copied into receive buffers

pub const fn tx_undersize_frames(&self) -> Option<u64>

The number of frames below the minimum length for the media. This would be less than 64 for Ethernet

pub const fn tx_oversize_frames(&self) -> Option<u64>

The number of frames longer than the maximum length for the media. This would be 1500 for Ethernet

pub const fn tx_dropped_frames(&self) -> Option<u64>

The number of valid frames that were dropped because received buffers were full

pub const fn tx_unicast_frames(&self) -> Option<u64>

The number of valid unicast frames transmitted and not dropped

pub const fn tx_broadcast_frames(&self) -> Option<u64>

The number of valid broadcast frames transmitted and not dropped

pub const fn tx_multicast_frames(&self) -> Option<u64>

The number of valid multicast frames transmitted and not dropped

pub const fn tx_crc_error_frames(&self) -> Option<u64>

The number of transmitted frames with CRC or alignment errors

pub const fn tx_total_bytes(&self) -> Option<u64>

The total number of bytes transmitted including error frames and dropped frames

pub const fn collisions(&self) -> Option<u64>

The number of collisions detected on this subnet

pub const fn unsupported_protocol(&self) -> Option<u64>

The number of frames destined for unsupported protocol

pub const fn rx_duplicated_frames(&self) -> Option<u64>

The number of valid frames received that were duplicated

pub const fn rx_decrypt_error_frames(&self) -> Option<u64>

The number of encrypted frames received that failed to decrypt

pub const fn tx_error_frames(&self) -> Option<u64>

The number of frames that failed to transmit after exceeding the retry limit

pub const fn tx_retry_frames(&self) -> Option<u64>

The number of frames that transmitted successfully after more than one attempt