uefi_raw::protocol

Module tcg

TCG (Trusted Computing Group) protocols.

These protocols provide access to the TPM (Trusted Platform Module).

There are two versions of the protocol. The original protocol is in the v1 module. It is used with TPM 1.1 and 1.2 devices. The newer protocol in the v2 module is generally provided for TPM 2.0 devices, although the spec indicates it can be used for older TPM versions as well.

Modules

v1

TCG (Trusted Computing Group) protocol for TPM (Trusted Platform Module) 1.1 and 1.2.

v2

TCG (Trusted Computing Group) protocol for TPM (Trusted Platform Module) 2.0.

Structs

AlgorithmId

Algorithm identifiers.

EventType

Event types stored in the TPM event log. The event type defines which structure type is stored in the event data.