uefi_raw::table

Module boot

UEFI services available during boot.

Structs

BootServices

Table of pointers to all the boot services.

EventType

Flags describing the type of an UEFI event and its attributes.

InterfaceType

Interface type of a protocol interface.

MemoryAttribute

Flags describing the capabilities of a memory range.

MemoryDescriptor

A structure describing a region of memory. This type corresponds to version of this struct in the UEFI spec and is always bound to a corresponding UEFI memory map.

MemoryType

The type of a memory range.

OpenProtocolInformationEntry

TimerDelay

Tpl

Task priority level.

Constants

PAGE_SIZE

Size in bytes of a UEFI page.

Type Aliases

EventNotifyFn

Raw event notification function.