Struct uefi_raw::table::boot::MemoryType

#[repr(transparent)]
pub struct MemoryType(pub u32);

The type of a memory range.

UEFI allows firmwares and operating systems to introduce new memory types in the 0x70000000..0xFFFFFFFF range. Therefore, we don’t know the full set of memory types at compile time, and it is not safe to model this C enum as a Rust enum.

Tuple Fields
0: u32
Implementations

impl MemoryType

pub const RESERVED: MemoryType = _

This enum variant is not used.

pub const LOADER_CODE: MemoryType = _

The code portions of a loaded UEFI application.

pub const LOADER_DATA: MemoryType = _

The data portions of a loaded UEFI applications, as well as any memory allocated by it.

pub const BOOT_SERVICES_CODE: MemoryType = _

Code of the boot drivers.

Can be reused after OS is loaded.

pub const BOOT_SERVICES_DATA: MemoryType = _

Memory used to store boot drivers’ data.

Can be reused after OS is loaded.

pub const RUNTIME_SERVICES_CODE: MemoryType = _

Runtime drivers’ code.

pub const RUNTIME_SERVICES_DATA: MemoryType = _

Runtime services’ code.

pub const CONVENTIONAL: MemoryType = _

Free usable memory.

pub const UNUSABLE: MemoryType = _

Memory in which errors have been detected.

pub const ACPI_RECLAIM: MemoryType = _

Memory that holds ACPI tables. Can be reclaimed after they are parsed.

pub const ACPI_NON_VOLATILE: MemoryType = _

Firmware-reserved addresses.

pub const MMIO: MemoryType = _

A region used for memory-mapped I/O.

pub const MMIO_PORT_SPACE: MemoryType = _

Address space used for memory-mapped port I/O.

pub const PAL_CODE: MemoryType = _

Address space which is part of the processor.

pub const PERSISTENT_MEMORY: MemoryType = _

Memory region which is usable and is also non-volatile.

impl MemoryType

pub const fn custom(value: u32) -> MemoryType

Construct a custom MemoryType. Values in the range 0x80000000..=0xffffffff are free for use if you are an OS loader.