Struct uefi_raw::time::Time

#[repr(C)]
pub struct Time {
    pub year: u16,
    pub month: u8,
    pub day: u8,
    pub hour: u8,
    pub minute: u8,
    pub second: u8,
    pub pad1: u8,
    pub nanosecond: u32,
    pub time_zone: i16,
    pub daylight: Daylight,
    pub pad2: u8,
}

Date and time representation.

Fields
year: u16

Year. Valid range: 1900..=9999.

month: u8

Month. Valid range: 1..=12.

day: u8

Day of the month. Valid range: 1..=31.

hour: u8

Hour. Valid range: 0..=23.

minute: u8

Minute. Valid range: 0..=59.

second: u8

Second. Valid range: 0..=59.

pad1: u8

Unused padding.

nanosecond: u32

Nanosececond. Valid range: 0..=999_999_999.

time_zone: i16

Offset in minutes from UTC. Valid range: -1440..=1440, or Time::UNSPECIFIED_TIMEZONE.

daylight: Daylight

Daylight savings time information.

pad2: u8

Unused padding.

Implementations

impl Time

pub const UNSPECIFIED_TIMEZONE: i16 = 2_047i16

Indicates the time should be interpreted as local time.

pub const fn invalid() -> Self

Create an invalid Time with all fields set to zero.

pub fn is_valid(&self) -> bool

True if all fields are within valid ranges, false otherwise.