uefi_raw::table::boot

Struct MemoryType

#[repr(transparent)]
pub struct MemoryType(pub u32);

The type of a memory range.

UEFI allows firmwares and operating systems to introduce new memory types in the 0x7000_0000..=0xFFFF_FFFF range. Therefore, we donโt know the full set of memory types at compile time, and it is not safe to model this C enum as a Rust enum.

Tuple Fields
0: u32
Implementations

impl MemoryType

pub const RESERVED: MemoryType

Not usable.

pub const LOADER_CODE: MemoryType

The code portions of a loaded UEFI application.

pub const LOADER_DATA: MemoryType

The data portions of a loaded UEFI applications, as well as any memory allocated by it.

pub const BOOT_SERVICES_CODE: MemoryType

Code of the boot drivers.

Can be reused after OS is loaded.

pub const BOOT_SERVICES_DATA: MemoryType

Memory used to store boot driversโ data.

Can be reused after OS is loaded.

pub const RUNTIME_SERVICES_CODE: MemoryType

Runtime driversโ code.

pub const RUNTIME_SERVICES_DATA: MemoryType

Runtime servicesโ code.

pub const CONVENTIONAL: MemoryType

Free usable memory.

pub const UNUSABLE: MemoryType

Memory in which errors have been detected.

pub const ACPI_RECLAIM: MemoryType

Memory that holds ACPI tables. Can be reclaimed after they are parsed.

pub const ACPI_NON_VOLATILE: MemoryType

Firmware-reserved addresses.

pub const MMIO: MemoryType

A region used for memory-mapped I/O.

pub const MMIO_PORT_SPACE: MemoryType

Address space used for memory-mapped port I/O.

pub const PAL_CODE: MemoryType

Address space which is part of the processor.

pub const PERSISTENT_MEMORY: MemoryType

Memory region which is usable and is also non-volatile.

pub const UNACCEPTED: MemoryType

Memory that must be accepted by the boot target before it can be used.

pub const MAX: MemoryType

End of the defined memory types. Higher values are possible though, see MemoryType::RESERVED_FOR_OEM and MemoryType::RESERVED_FOR_OS_LOADER.

impl MemoryType

pub const RESERVED_FOR_OEM: RangeInclusive<u32>

Range reserved for OEM use.

pub const RESERVED_FOR_OS_LOADER: RangeInclusive<u32>

Range reserved for OS loaders.

pub const fn custom(value: u32) -> Self

Construct a custom MemoryType. Values in the range 0x8000_0000..=0xffff_ffff are free for use if you are an OS loader.