uefi_raw

Union IpAddress

#[repr(C)]
pub union IpAddress {
    pub addr: [u32; 4],
    pub v4: Ipv4Address,
    pub v6: Ipv6Address,
}

An IPv4 or IPv6 internet protocol address.

Corresponds to the EFI_IP_ADDRESS type in the UEFI specification. This type is defined in the same way as edk2 for compatibility with C code. Note that this is an untagged union, so thereโs no way to tell which type of address an IpAddress value contains without additional context.

Fields
addr: [u32; 4]

This member serves to align the whole type to a 4 bytes as required by the spec. Note that this is slightly different from repr(align(4)), which would prevent placing this type in a packed structure.

v4: Ipv4Address

An IPv4 internet protocol address.

v6: Ipv6Address

An IPv6 internet protocol address.

Implementations

impl IpAddress

pub const fn new_v4(ip_addr: [u8; 4]) -> Self

Construct a new IPv4 address.

pub const fn new_v6(ip_addr: [u8; 16]) -> Self

Construct a new IPv6 address.